+++ docs/release_notes.rst
@@ -4,7 +4,8 @@ Release Notes
 1.0b2
 -----
 
- * Adds Django 1.11 to test matrix (Benjamin Bach) #31
+ * Correctly set default options for ``notifymail`` management command (Benjamin Bach) #32
+ * Adds Django 1.11 to test matrix (Benjamin Bach) #32
  * Do not return ``bytes`` in ``__str__`` (Øystein Hiåsen) #28
 
 1.0b1